pes-header PTS/DTS clock parsing: fix 2 bugs in value expansion logic. 1) lastByte assignment has to be priorized properly within wrapping operation 2) byte 14 instead of 13 was read whilst not using lastByte reg.
Thus the wrong LSB was read for PTS = DTS, but also reading 1 byte more than we should, and finally causing the read offset for PTS != DTS to be 1 byte forward than expected. that extra offset was however not causing transport level packet skip issues as the remainder length was and is computed invariable to the number of read bytes (see end of func).